requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
crispy-bootstrap5==2024.10
cryptography==44.0.2
cssselect2==0.8.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.2.20
django-allauth==65.4.0
django-crispy-forms==2.3
django-filter==24.3
django-geojson==4.1.0
django-guardian==2.4.0
django-import-export==4.3.5
django-leaflet==0.31.0
django-mathfilters==1.0.0
django-phonenumber-field==8.0.0
django-simple-history==3.8.0
django-tables2==2.7.5
django-tinymce==4.1.0
django-unicorn==0.61.0
djangorestframework==3.15.2
Faker==37.0.0
fonttools==4.56.0
html5lib==1.1
idna==3.10
lxml==5.3.1
orjson==3.10.15
oscrypto==1.3.0
phonenumbers==8.13.54
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.7
pypdf==5.3.1
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.0
requests==2.32.3
shortuuid==1.0.13
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
gunicorn


tqdm~=4.67.1
python-dotenv~=1.1.0